No more equal keys :( But more files to structure :)
